requirements.txt
playwright>=1.58.0
httpx>=0.28.1
PyMuPDF>=1.27.1
beautifulsoup4>=4.12.2
google-genai>=1.65.0
openai>=2.24.0
python-dotenv>=1.2.1
sentence-transformers>=5.2.3
fastapi>=0.134.0
uvicorn>=0.41.0
requests>=2.32.5
sqlalchemy>=2.0.47
aiosqlite>=0.22.1
pydantic-settings>=2.13.1
chromadb>=1.5.2
pytest>=9.0.2
